[bug] 할당되지 않은 key, value 추가
