+++ ds_envp/cast_arr.c
@@ -6,7 +6,7 @@
 /*   By: jeongmin <jeongmin@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/07 17:17:28 by jeongmin          #+#    #+#             */
-/*   Updated: 2023/01/08 17:02:21 by jeongmin         ###   ########.fr       */
+/*   Updated: 2023/01/08 22:55:40 by jeongmin         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
